// frontend/src/api/types.ts
export type LedgerType =
  | 'P2P_TRANSFER' | 'CASH_IN' | 'PIX_INTERNAL' | 'PIX_OUT' | 'PIX_OUT_REVERSAL' | 'PIX_IN' | 'CHARGE_PAYMENT'
  | 'BILL_PAYMENT' | 'BILL_PAYMENT_REVERSAL' | 'YIELD_CREDIT' | 'LOAN_DISBURSEMENT' | 'LOAN_INSTALLMENT_PAYMENT'
  | 'CARD_HOLD' | 'CARD_HOLD_RELEASE' | 'CARD_DEBIT_CLEARING' | 'CARD_CREDIT_CLEARING' | 'CARD_REVOLVING_INTEREST'
  | 'CARD_STATEMENT_PAYMENT' | 'MARKETPLACE_PURCHASE' | 'MARKETPLACE_REFUND' | 'MARKETPLACE_COMMISSION'
  | 'CASHBACK_CREDIT' | 'PIX_RETURN' | 'PIX_RETURN_REVERSAL' | 'PIX_MED_BLOCK' | 'PIX_MED_RELEASE' | 'PIX_MED_RETURN'
  | 'CHARGE_REFUND' | 'CARD_REFUND' | 'CARD_CHARGEBACK' | 'LOAN_PREPAYMENT' | 'CARD_STATEMENT_FINANCING'